requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.2.0=pypi_0
aiohttp=3.7.4.post0=pypi_0
alembic=1.8.1=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
arrow=1.2.2=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=3.0.1=pypi_0
attrs=22.1.0=pypi_0
autopage=0.5.1=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
boto3=1.24.51=pypi_0
botocore=1.27.51=pypi_0
bottleneck=1.3.5=py38h7deecbd_0
bravado=11.0.3=pypi_0
bravado-core=5.17.0=pypi_0
brotlipy=0.7.0=py38h0a891b7_1004
ca-certificates=2023.7.22=hbcca054_0
cachetools=5.2.0=pypi_0
captum=0.4.0=pypi_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py38h74dc2b5_0
chardet=4.0.0=pypi_0
charset-normalizer=2.1.0=pypi_0
click=8.1.3=unix_pyhd8ed1ab_2
cliff=4.0.0=pypi_0
cmaes=0.8.2=pypi_0
cmd2=2.4.2=pypi_0
colorama=0.4.5=pyhd8ed1ab_0
colorlog=6.6.0=pypi_0
commonmark=0.9.1=pypi_0
cryptography=35.0.0=py38h3e25421_2
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.11.0=pyhd8ed1ab_0
datasets=1.2.1=pypi_0
decorator=5.1.1=pyhd3eb1b0_0
deepspeed=0.7.3=pypi_0
dill=0.3.5.1=pypi_0
dython=0.6.7.post2=pyhd8ed1ab_0
et_xmlfile=1.0.1=py_1001
executing=0.8.3=pyhd3eb1b0_0
fftw=3.3.10=nompi_h77c792f_102
filelock=3.8.0=pyhd8ed1ab_0
fonttools=4.34.4=pypi_0
fqdn=1.5.1=pypi_0
freetype=2.11.0=h70c0345_0
fsspec=2021.10.0=pypi_0
future=0.18.2=pypi_0
giflib=5.2.1=h7b6447c_0
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
google-auth=2.10.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
greenlet=1.1.2=pypi_0
grpcio=1.47.0=pypi_0
hjson=3.1.0=pypi_0
huggingface-hub=0.8.1=pypi_0
huggingface_hub=0.11.0=pyhd8ed1ab_0
hydra=2.5=pypi_0
hydra-colorlog=1.1.0=pypi_0
hydra-core=1.2.0=pypi_0
hydra-optuna-sweeper=1.1.0=pypi_0
idna=3.3=pypi_0
importlib-metadata=4.12.0=pypi_0
importlib-resources=5.9.0=pypi_0
importlib_metadata=5.0.0=hd8ed1ab_1
intel-openmp=2021.4.0=h06a4308_3561
ipython=8.4.0=py38h06a4308_0
isoduration=20.11.0=pypi_0
jedi=0.18.1=py38h06a4308_1
jmespath=1.0.1=pypi_0
joblib=1.1.0=pypi_0
jpeg=9b=h024ee3a_2
jsonlines=1.2.0=pypi_0
jsonpointer=2.3=pypi_0
jsonref=0.2=pypi_0
jsonschema=4.9.1=pypi_0
kiwisolver=1.4.4=py38h6a678d5_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libgomp=11.2.0=h1234567_1
libllvm10=10.0.1=he513fc3_3
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.1.0=h2733197_1
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
llvmlite=0.36.0=py38h4630a5e_0
lz4-c=1.9.3=h295c915_1
mako=1.2.1=pypi_0
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.3=pypi_0
matplotlib-base=3.4.3=py38hf4fb855_1
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
monotonic=1.6=pypi_0
msgpack=1.0.4=pypi_0
multidict=5.2.0=pypi_0
multiprocess=0.70.12.2=pypi_0
ncurses=6.3=h5eee18b_3
neptune-client=0.14.2=pypi_0
ninja=1.10.2.4=pypi_0
ninja-base=1.10.2=hd09550d_5
nltk=3.7=pyhd8ed1ab_0
numba=0.53.1=py38ha9443f7_0
numexpr=2.8.3=py38h807cd23_0
numpy=1.19.5=pypi_0
oauthlib=3.2.0=pypi_0
omegaconf=2.2.3=pypi_0
openai=0.23.1=pyhd8ed1ab_0
openpyxl=3.0.9=pyhd8ed1ab_0
openssl=1.1.1o=h166bdaf_0
optuna=2.4.0=pypi_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.3=pypi_0
pandas-stubs=1.5.0.221012=pyhd8ed1ab_0
parso=0.8.3=pyhd3eb1b0_0
pbr=5.10.0=pypi_0
pexpect=4.8.0=pyhd3eb1b0_3
pickle5=0.0.11=pypi_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py38hace64e9_1
pip=22.2.2=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
prettytable=3.3.0=pypi_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.19.4=pypi_0
psutil=5.9.1=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
py-cpuinfo=8.0.0=pypi_0
pyarrow=5.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pydantic=1.10.2=pypi_0
pydeprecate=0.3.1=pypi_0
pygments=2.12.0=pypi_0
pyjwt=2.4.0=pypi_0
pynndescent=0.5.10=pyh1a96a4e_0
pyopenssl=22.0.0=pyhd8ed1ab_1
pyparsing=3.0.9=pyhd8ed1ab_0
pyperclip=1.8.2=pypi_0
pyrsistent=0.18.1=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.8.11=h12debd9_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.7.1=py3.8_cuda10.1.243_cudnn7.6.3_0
pytorch-lightning=1.4.8=pypi_0
pytz=2022.2.1=pypi_0
pyyaml=6.0=py38h0a891b7_4
readline=8.1.2=h7f8727e_1
regex=2022.7.25=pypi_0
requests=2.28.1=pyhd8ed1ab_1
requests-oauthlib=1.3.1=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3987=1.3.8=pypi_0
rich=10.6.0=pypi_0
rsa=4.9=pypi_0
s3transfer=0.6.0=pypi_0
scikit-learn=1.0.2=pypi_0
scikit-plot=0.3.7=py_1
scipy=1.9.0=pypi_0
seaborn=0.12.2=py38h06a4308_0
sentence-transformers=2.2.2=pyhd8ed1ab_0
sentencepiece=0.1.96=pypi_0
setuptools=63.4.1=py38h06a4308_0
simplejson=3.17.6=pypi_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sqlalchemy=1.4.40=pypi_0
sqlite=3.39.3=h5082296_0
stack_data=0.2.0=pyhd3eb1b0_0
stevedore=4.0.0=pypi_0
swagger-spec-validator=2.7.4=pypi_0
tbb=2020.2=h4bd325d_4
tensorboard=2.10.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.12.1=pypi_0
torchaudio=0.7.2=py38
torchmetrics=0.7.2=pypi_0
torchvision=0.8.2=py38_cu101
tornado=6.1=py38h0a891b7_3
tqdm=4.48.2=pypi_0
traitlets=5.1.1=pyhd3eb1b0_0
transformers=4.21.1=pypi_0
types-pytz=2022.5.0.0=pyhd8ed1ab_0
typing-extensions=4.3.0=py38h06a4308_0
typing_extensions=4.3.0=py38h06a4308_0
ujson=5.6.0=pypi_0
umap-learn=0.5.3=py38h578d9bd_1
uri-template=1.2.0=pypi_0
urllib3=1.26.11=pyhd8ed1ab_0
wcwidth=0.2.5=pyhd3eb1b0_0
webcolors=1.12=pypi_0
websocket-client=1.3.3=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xxhash=2.0.2=pypi_0
xz=5.2.6=h5eee18b_0
yaml=0.2.5=h7f98852_2
yarl=1.7.0=pypi_0
zipp=3.8.1=pypi_0
zlib=1.2.12=h5eee18b_3
zstd=1.4.9=haebb681_0

